extract all files copy specific files not working (#1966)
when using having `to: "kubect-rook-ceph.sh"` in files section
it was not working for us but it worked in previous release.

```kubectl rook-ceph
Error: unknown command "rook-ceph" for "kubectl"
Run 'kubectl --help' for usage.
```

but modifying to `to: "."` is working

```
kubectl rook-ceph

DESCRIPTION
kubectl rook-ceph provides common management and troubleshooting tools for Ceph.
....
```

Signed-off-by: subhamkrai <srai@redhat.com>